@@ -1,26 +1,27 @@
 module github.com/elastic/integrations
 
-go 1.21.0
+go 1.22.0
 
 require (
 	github.com/blang/semver v3.5.1+incompatible
-	github.com/elastic/elastic-package v0.96.1
-	github.com/elastic/go-licenser v0.4.1
-	github.com/elastic/package-registry v1.23.0
+	github.com/cli/go-gh/v2 v2.9.0
+	github.com/elastic/elastic-package v0.101.2
+	github.com/elastic/go-licenser v0.4.2
+	github.com/elastic/package-registry v1.24.0
 	github.com/magefile/mage v1.15.0
 	github.com/pkg/errors v0.9.1
-	github.com/stretchr/testify v1.8.4
-	golang.org/x/tools v0.17.0
-	gopkg.in/yaml.v2 v2.4.0
+	github.com/stretchr/testify v1.9.0
+	golang.org/x/tools v0.23.0
+	gopkg.in/yaml.v3 v3.0.1
 )
 
 require (
-	cloud.google.com/go v0.110.10 // indirect
-	cloud.google.com/go/compute v1.23.3 // indirect
+	cloud.google.com/go v0.112.1 // indirect
+	cloud.google.com/go/compute v1.24.0 // indirect
 	cloud.google.com/go/compute/metadata v0.2.3 // indirect
-	cloud.google.com/go/iam v1.1.5 // indirect
-	cloud.google.com/go/pubsub v1.33.0 // indirect
-	cloud.google.com/go/storage v1.35.1 // indirect
+	cloud.google.com/go/iam v1.1.7 // indirect
+	cloud.google.com/go/pubsub v1.37.0 // indirect
+	cloud.google.com/go/storage v1.40.0 // indirect
 	github.com/AlecAivazis/survey/v2 v2.3.7 // indirect
 	github.com/Azure/go-ansiterm v0.0.0-20210617225240-d185dfc1b5a1 // indirect
 	github.com/MakeNowJust/heredoc v1.0.0 // indirect
