@@ -18,7 +18,9 @@ def find_oldest_supported_version(kibana_version_condition: str) -> str:
     if "||" in kibana_version_condition and kibana_version_condition.index("||") >= 0:
         return handle_or(kibana_version_condition)
 
-    available_versions = fetch_version()
+    artifacts_versions = fetch_version()
+    available_versions = artifacts_versions.get("versions", [])
+    available_aliases = artifacts_versions.get("aliases", [])
     version = remove_operator(kibana_version_condition)
     parts = version.split(".")
 
@@ -32,22 +34,22 @@ def find_oldest_supported_version(kibana_version_condition: str) -> str:
     # Use the snapshot if this is the last patch version.
     next_patch = ".".join((major, minor, str(int(patch)+1)))
     next_patch_exists = (
-        next_patch in available_versions.get("versions", []) or
-        f"{next_patch}-SNAPSHOT" in available_versions.get("versions", [])
+        next_patch in available_versions or
+        f"{next_patch}-SNAPSHOT" in available_versions
     )
 
     snapshot_version = f"{version}-SNAPSHOT"
-    if not next_patch_exists and (snapshot_version in available_versions.get("versions", [])):
+    if not next_patch_exists and (snapshot_version in available_versions):
         return snapshot_version
 
     # Use the version as is if it exists.
-    if version in available_versions.get("version", []):
+    if version in available_versions:
         return version
 
     # Old minors may not be available in artifacts-api, if it is older
     # than the others in the same major, return the version as is.
     older = True
-    for available_version in available_versions.get("versions", []):
+    for available_version in available_versions:
         available_parts = available_version.split(".")
         if len(available_parts) < 2:
             continue
@@ -63,7 +65,7 @@ def find_oldest_supported_version(kibana_version_condition: str) -> str:
     # If no version has been found so far, try with the snapshot of the next version
     # in the current major.
     major_snapshot = f"{major}.x-SNAPSHOT"
-    if major_snapshot in available_versions.get("aliases", []):
+    if major_snapshot in available_aliases:
         return major_snapshot
 
     # Otherwise, return it, whatever this is.
